[Controls] FilterComboBox: Correctly cancel selection or edition
Up until this commit, any action causing the filtering text field to lose
the focus was considered to be a valid edition, thus propagating the value
in the text field or the one highlighted to be propagated, even if the
user was clicking anywhere else in the application, or pressing the
`esc` key.

The distinction is now made between cases where the edition is finished
because it was validated (value clicked on, or selected with the `return`/
`enter` key), and not because it was cancelled (clicks outside of the
combobox or `esc` key pressed).