[ui] Correctly distinguish "save" from "save as" cases in `unsavedDialog`
Prior to this commit, even if the current graph had never been saved,
we ended up going through the "save current file" use case rather than
the "save as" because the `saveAction` button was always enabled.

This led to the callback being fired before the "save as" dialog had
been closed.