[core] Make `range` optional for `IntParam` and `FloatParam`
`range` is set to `None` by default, making it optional. If it is
specified, correct values are expected (e.g. only floats for `FloatParam`)
 and a slider will be displayed in the Node Editor. However, if it is not
specified, no slider will be displayed and the value will be set at the
discretion of the user.

Having `range=None` will not trigger any "invalid" error in the
description. It is thus not mandatory anymore to specify the valid range
of values for the integer and floating attributes in node descriptions.