[ui] Node: Implementation of the Backdrop Node interface.
A Backdrop node is default of pale yellow colour and is of the same colour in the header and the bottom area representing the node. Backdrop node allows user resize by dragging on the mouse area at the edges and also allows text appearing below the header from the notes section in the attribute panel.