[qt6][Viewer] Viewer2D: Fix `fit()` with the Sequence Player
With the Sequence Player always enabled by default, the `fit()` function
must be triggered at the right moment, and only once the layout has
been correctly initialized; otherwise, zoom factors may become negative
and the images will not be placed correctly in the viewer.

Changes in the orientation tag must be handled independently from the
state of the zoom factor.