[core] Node: Add a `dynamicEnvVars` dictionary as a node member
`dynamicEnvVars` is a dictionary currently mapping a name to a folder
to use as the root when writing relative paths. Keys correspond to
the variable names that are to be replaced with their values during
the evaluation of attributes.

Although it currently only contains folder paths, any type of variable
may be written in this dictionary to resolve attribute values with.