[tests] Update `test_multiviewPipeline` with test nodes and template
`test_multiviewPipeline` relied on the photogrammetry template and its
nodes, which are not a part of Meshroom anymore.

The test is rewritten to check the same items (serialization, graph
equality, etc.) but using dedicated test nodes (which already existed)
and template, which is added with this commit.

The path of the template is added to `MESHROOM_PIPELINE_TEMPLATES_PATH`
when loading the test module.