[ui] Compatibility nodes cannot be computed or recomputed
When the compatibility node was fully computed, the recompute/resubmit
button was available in the UI menu.
The reason is that we need to propagate that the connected nodes can be
computed if there is a compatibility node in the dependency iff it is
already fully computed.
Now there is an additional explicit check.