[ui] Add a "Reload All Nodes" menu to reload all the plugins' nodes
"Reload All Nodes" will reload all the descriptions for `NodePlugins`
that are part of loaded plugins, whether they are registered or not.

Once the reload of the descriptions has been performed, the node instances
from the current graph will be reloaded as well to reflect any change
in the description.