[tests] Plugins: Add a `sleep` between file rewrites
On some systems such as GitHub's Windows CI, the `write` operation
is too fast and does not cause a change in the timestamp of the
file we're reloading, hence causing the test to fail for external
reasons.

Adding a sleep does not change anything to the test functionally, but
on the contrary ensures that we are actually testing the feature.

https://stackoverflow.com/questions/19059877/python-os-path-getmtime-time-not-changing