`reloadAllNodes`: Save reloaded types and only replace targeted nodes
When requesting a `NodePlugin` to reload, the status of the reloading
is now saved. If it has been successful, all the nodes of this type in
the graph will effectively be removed. If it has not (either because of
an error, or because no change has been done to the description), then
nothing will happen.

Replacing only the nodes whose description has been updated prevents
from having to update heavy graphs for very few nodes to change (if any).