[core] DirTreeProcessEnv: add internal site-packages first
Ensures that the packages installed in a local virtualenv is taken
before the propagation of the global Meshroom environment to avoid
version conflicts.

Note: We tried to avoid checking files/folders on disk at startup for
performance reasons, but here we are forced to use glob as there is no
way to know the python version of the plugin in advance.