[core] plugins: Load plugin's configuration file upon its initialisation
When initialising a plugin, Meshroom now looks for a `config.json` file
that contains the environment variables that need to be set for it to
work properly.

If such a file is found, the environment variables and their values
are stored by the plugin, and will be used when resolving the
comamnd line.

If variables such as paths to files or folders need to
be set, they are checked for and converted to absolute paths.

If no such file exists, nothing happens.