[Controls] Add minimum width for `MessageDialogs`
In cases where the title of the dialog is short (for example when the
version is frozen into release), setting a minimum width ensures that
the dialogs are wide enough to remain readable.

If the title of the dialog is greater than the set minimum width, then
this size will be preferred over the minimum one.