[ImageGallery] Remove useless `ColumnLayout`
Following the removal of the "Augment Reconstruction" area, there is no
use for a `ColumnLayout` element, as the label to drop the images can be
used on its own.

This commit also removes the check on the hovering which determined
whether the mouse was over the "Drop images" or "Augment reconstruction"
area as only the former remains.