[core] Split between cmdVars, staticCmdVars and command line vars
- Command line vars are now only computed during the command line
generation for the "CommandLineNode".
- internalFolder is now initialized in the BaseNode, instead of Node
(even if there is the particular case of "InputNode" which does not need
it). And its value is computed once in the updateInternals instead of
redoing the string format again and again.
- internalFolder is not serialized/deserialized anymore as it is a
static information. At the beginning, it was designed to be customized
per node, but this flexibility/complexity seems useless as it has never
been used over many years. We now use "{nodeCacheFolder}" directly.