[core] Rez plugins: Ensure dependencies for `meshroom_compute` are present
`psutil` is a dependency of `meshroom_compute`, which cannot run if it
is not part of the environment. For rez plugins, if `psutil` is not
already part of the plugin's environment, it is added. No version is
specified so it can be resolved in the environment, without being aligned
on Meshroom's.