[core] Split version name on the correct separator
In adbdd3b99f539c77699a32fa138e479f3b275f10, the version name for
develop versions has been updated to use the "+" separator instead of
"-". When loading a scene that has been saved since this change,
errors occur during the parsing of the version name as the "+" is not
expected. This commit fixes this issue by declaring "+" as the separator.

Scenes that were saved with "-" remain compatible, as we replace all
"-" in the version name with "+" before splitting it.