[qml] ExpressionTextField: Don't round up floating numbers
If the `ExpressionTextField` is meant for an integer (`isInt = true`),
then the value of the expression should be an integer, meaning rounding
may need to be applied. However, for floating numbers, we do not want
to lose some precision, so the value should never be rounded or set
to a fixed number of decimal digits.