[core] node: Add `chunkPlaceholder` property
`chunkPlaceholder` is a list model that is meant to contain a placeholder
chunk for nodes that have dynamic chunks: prior to the chunks' creation,
it is used to reflect the state of the node as a single chunk.

In particular, this allows to reflect accurately the node's status while
it is submitted but the chunks have not been created yet.

The property is a list model as it allows it to be used as a model later
on on the QML side.