[core] node: Add retro-compatibility for status files of single-chunk nodes
Following some recent changes, all the status/log/statistics files are
always prefixed by the chunk number, if the node only has a single chunk.
Before this, nodes with a single chunk had files that were not prefixed.

With this commit, it is possible to open a scene that contains nodes
(with or without several chunks) that were computed before these changes
were introduced without losing track of these files. Without it, all the
nodes that have a single chunk appear as though they have never been
computed, even if their cache exists on disk.